+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Ductwork Changelog
 
+## Unreleased
+
+- fix: allow steps to be chained while pipeline is expanded or divided (before collapsing or combining) - before this incorrectly raised a `CollapseError` or `CombineError`
+
 ## [0.1.0]
 
 - Initial release - see [documentation](https://docs.getductwork.io/) for details